Remove `orion_utils` module
