+++ src/prefect/cli/profile.py
@@ -5,7 +5,9 @@ import os
 import textwrap
 from typing import Optional
 
+import httpx
 import typer
+from fastapi import status
 from rich.progress import Progress, SpinnerColumn, TextColumn
 from rich.table import Table
 
@@ -13,9 +15,11 @@ import prefect.context
 import prefect.settings
 from prefect.cli._types import PrefectTyper
 from prefect.cli._utilities import exit_with_error, exit_with_success
-from prefect.cli.orion_utils import ConnectionStatus, check_orion_connection
+from prefect.cli.cloud import CloudUnauthorizedError, get_cloud_client
 from prefect.cli.root import app
+from prefect.client import get_client
 from prefect.context import use_profile
+from prefect.utilities.collections import AutoEnum
 
 profile_app = PrefectTyper(
     name="profile", help="Commands for interacting with your Prefect profiles."
@@ -146,7 +150,7 @@ async def use(name: str):
     ) as progress:
 
         progress.add_task(
-            description="Connecting...",
+            description="Checking API connectivity...",
             total=None,
         )
 
@@ -238,3 +242,65 @@ def inspect(
 
     for setting, value in profiles[name].settings.items():
         app.console.print(f"{setting.name}='{value}'")
+
+
+class ConnectionStatus(AutoEnum):
+    CLOUD_CONNECTED = AutoEnum.auto()
+    CLOUD_ERROR = AutoEnum.auto()
+    CLOUD_UNAUTHORIZED = AutoEnum.auto()
+    ORION_CONNECTED = AutoEnum.auto()
+    ORION_ERROR = AutoEnum.auto()
+    EPHEMERAL = AutoEnum.auto()
+    INVALID_API = AutoEnum.auto()
+
+
+async def check_orion_connection():
+    httpx_settings = dict(timeout=3)
+    try:
+        # attempt to infer Cloud 2.0 API from the connection URL
+        cloud_client = get_cloud_client(
+            httpx_settings=httpx_settings, infer_cloud_url=True
+        )
+        await cloud_client.api_healthcheck()
+        return ConnectionStatus.CLOUD_CONNECTED
+    except CloudUnauthorizedError:
+        # if the Cloud 2.0 API exists and fails to authenticate, notify the user
+        return ConnectionStatus.CLOUD_UNAUTHORIZED
+    except httpx.HTTPStatusError as exc:
+        if exc.response.status_code == status.HTTP_404_NOT_FOUND:
+            # if the route does not exist, attmpt to connect as a hosted Orion instance
+            try:
+                # inform the user if Prefect Orion endpoints exist, but there are
+                # connection issues
+                client = get_client(httpx_settings=httpx_settings)
+                connect_error = await client.api_healthcheck()
+                if connect_error is not None:
+                    return ConnectionStatus.ORION_ERROR
+                elif await client.using_ephemeral_app():
+                    # if the client is using an ephemeral Orion app, inform the user
+                    return ConnectionStatus.EPHEMERAL
+                else:
+                    return ConnectionStatus.ORION_CONNECTED
+            except Exception as exc:
+                return ConnectionStatus.ORION_ERROR
+        else:
+            return ConnectionStatus.CLOUD_ERROR
+    except TypeError:
+        # if no Prefect Orion API URL has been set, httpx will throw a TypeError
+        try:
+            # try to connect with the client anyway, it will likely use an
+            # ephemeral Orion instance
+            client = get_client(httpx_settings=httpx_settings)
+            connect_error = await client.api_healthcheck()
+            if connect_error is not None:
+                return ConnectionStatus.ORION_ERROR
+            elif await client.using_ephemeral_app():
+                return ConnectionStatus.EPHEMERAL
+            else:
+                return ConnectionStatus.ORION_CONNECTED
+        except Exception as exc:
+            return ConnectionStatus.ORION_ERROR
+    except (httpx.ConnectError, httpx.UnsupportedProtocol) as exc:
+        return ConnectionStatus.INVALID_API
+
+    return exit_method, msg
